Refactor: move Kernel method implementations from object.rs to kernel.rs
object.rs now contains only BasicObject methods (initialize, ==, equal?,
!, !=, __id__, __send__, instance_eval, instance_exec, method_missing,
singleton_method_added/removed/undefined).

All Kernel methods (class, hash, eql?, dup, to_s, inspect, respond_to?,
is_a?, instance_of?, send, object_id, method, methods, singleton_methods,
singleton_class, extend, enum_for, define_singleton_method,
instance_variable_*, initialize_copy/clone/dup, etc.) now live in
kernel.rs where they belong, matching CRuby's architecture.
